Complete dynamic rewrite, using _arguments
Everything is now generated dynamically from the output of "virsh help"
and "virsh help $cmd", except for the information which kind of
completion which command uses ("start needs an inactive domain").

Previously, commands could have occured twice if their description in
`virsh help` didn't match with the one contained in the completion file
(for example, translations). This can no longer happen, as they are now
taken from `virsh help` even if more information is available for that
command.

Instead of the cryptic regex mechanisms, we now use the normal
_arguments and _describe functions.

Parsing `virsh help $cmd` for information about whether a domain or a
file is expected for commands where this information isn't contained in
the completion is not yet implemented. This would be a great feature
however, even though this detection couldn't decide if any domain will
do or if for example it must be running.